MFS Gone GOLD
This will officially mark Medius File Services as GOLD.
RT packet System_Message now has proper implemented enums.
MediusFileUploadRequest now uploads properly with packet struct.
Simplified some handshake coding.

+++ Server.Dme/DMEMediusManager.cs
@@ -355,7 +355,7 @@ namespace Server.Dme
                                 ListenServerAddress = new NetAddress()
                                 {
                                     Address = Program.SERVER_IP.ToString(),
-                                    Port = (uint)Program.TcpServer.Port
+                                    Port = Program.TcpServer.Port
                                 }
                             }
                         });
